// src/api/models/bpmn-diagram-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface BpmnDiagramTO
 */
export interface BpmnDiagramTO {
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramTO
     */
    bpmnDiagramId?: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramTO
     */
    bpmnRepositoryId: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramTO
     */
    bpmnDiagramName: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramTO
     */
    bpmnDiagramDescription: any;
    /**
     * 
     * @type {Date}
     * @memberof BpmnDiagramTO
     */
    createdDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof BpmnDiagramTO
     */
    updatedDate?: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramTO
     */
    svgPreview?: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramTO
     */
    fileType?: any;
}