// src/api/models/bpmn-repository-request-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface BpmnRepositoryRequestTO
 */
export interface BpmnRepositoryRequestTO {
    /**
     * 
     * @type {string}
     * @memberof BpmnRepositoryRequestTO
     */
    bpmnRepositoryId: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnRepositoryRequestTO
     */
    bpmnRepositoryName: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnRepositoryRequestTO
     */
    bpmnRepositoryDescription: any;
    /**
     * 
     * @type {number}
     * @memberof BpmnRepositoryRequestTO
     */
    existingDiagrams: any;
    /**
     * 
     * @type {number}
     * @memberof BpmnRepositoryRequestTO
     */
    assignedUsers: any;
}